Fix segfault while running tests on macOS
The tests run in a different order on macOS than Linux, which happens to
result in this segfault.

The problem is that `test_compressed_grid` leaves some stuff behind in
`mutable_device().logical_block_types`, which have `pb_type` set to
`nullptr`. This then breaks `test_connection_router`'s call to
`XmlReadArch`, since it assumes that all of the logical block types are
ones that it created, all of which either have `pb_type` set or are
empty.

Making `test_compressed_grid` clear `logical_block_types` when it's done
fixes that.