[APPack] Added read_flat_place Option and Verifier
Added the read_flat_place option to VPR's CLI. This allows the user to
pass a flat placement file into the packer.

Added code to parse this file into a FlatPlacementInfo object which only
verifies that the atom given in the file exists in the atom netlist.

Added code in the packer to read in the flat placement and verify that
the flat placement is valid for use in the packer.