[APPack] Added Flat Placement Reconstruction
Added logic to the packer to reconstruct clusters from the flat
placement, if one is provided. The packer will fall-back on the original
packing algorithm if it runs out of options for candidates to cluster.

Added logic to the initial placer to reconstruct the placement of each
cluster from the flat placement, if one is provided. It will try to
place the cluster at the centroid location of all atoms in a cluster, if
it cannot place the cluster there it will fall-back on the original
initial placement algorithm.

Added a simple testcase to ensure the reading and writing of the flat
placement file is working.