[APPack] Added Max Displacement Metric
Added the ability to print the max displacement from the placement
reconstruction to the original flat placement.

This metric will help show another dimension to the flat placement
reconstruction. Generally, if the average displacement is low, that may
not be good if that means that one atom is placed way farther away than
where it would like. This can really hurt quality.

Turned the distance calculation into the Manhattan Norm instead of the
Euclidean Norm.