[CI] Added System Verilog Regression Tests to GitHub Runners
The system verilog regression tests were put into the currently
not-working nightly tests. Since they are so small, putting them in the
GitHub runners tests instead so they do not regress.

The strong parmys test also has system verilog tests (I think); however,
that test may be too large for the GitHub runners. We may consider
moving that test to the SAVI workflows in the future.