[Prepack] Documentation on Molecule Chains
Improved the documentation on how chains are handled in the prepacker.
The prepacker segments chains into multiple molecules (where each
molecule can fit into a single cluster). The prepacker maintains
MoleculeChainIds to the chain these molecules are a part of. This allows
the cluster legalizer to easily put the molecules into clusters and
understand when a molecule is part of a special chain (for gain
calculation).