[Pack] Removed Global Overfilled Partition Regions
After each pack, the partition regions are checked to see if any of the
partition regions are overfilled. During this process, the overfilled
partition regions were stored in the Floorplanning context (likely for
legacy reasons). This was then read from a few lines later in another
function.

Turned this into a local variable instead of a global variable.