[AP][PartialLegalizer] Density Abstraction
Abstracted how density is managed in the AP flow into a manager class.
This manager class contains two sub-classes: one for the bins and one
for the mass calculation.

Density is a function of mass and volume.

The volume is related to how we break the FPGA chip into discrete bins.
These bins contain atoms and they take up a rectangular amount of space
on the FPGA.

The mass is related to how "heavy" each APBlock is inside of the bin.
This is computed using an M-dimensional vector where M is the number of
models in the architecture. Each logical / physical tile has a capacity
that it may hold.

The density manager class manages how the bins are constructed, how much
mass each bin can contain, and how much mass each bin currently
contains. This current version of the density manager class creates a
bin for each tile (with a unique root tile location). The capacity of
each bin is related to the capacity of the physical tile at that
location.

Partial Legalizers can ask questions to the density manager class. They
may ask for all of the overfilled bins, what bin is located at a given
location, what bin contains a given atom, how overfilled / underfilled a
bin is, etc.

This abstraction will make it much easier to optimize the density of the
flat placement, since it separates algorithm from denisty calculation.
This will make writing future Partial Legalizers considerably easier.