[Place] Cleaned Up Placement Context Initialization
Cleaned up how the variables in the placement context were initialized
to make it easier to initialize the placement context outside of the
placement stage (Analytical Placement).

Added an optional argument to the Placer class to allow an initial
placement to be passed in. This allows a different initial placement to
be used without having to pull the initial placer out of the Placer
class.

With this change, it should be easier to create the Detailed Placer in
Analytical Placement.