[CI] Made Nightly Tests Continue on Error
Currently, the Nightly Tests will stop whenever any of the Nightly Tests
fail (and will not run the following tests). So for example, if Nightly
Test 1 has a QoR failure, all other Nightly tests will not be run.

Made each step in the Nightly Test manual job continue on error. This
will run the following steps even if one of them fail.