[AP][DetailedPlacer] Added Detailed Placer Class
The Detailed Placer is a pass in the AP flow which takes a legal
placement and optimizes it, while still remaining legal.

Previously in the AP flow, the annealer from the Placement stage was
always used to optimize the placement. Now the user can choose to not
use the annealer, by selecting the "none" Detailed Placer. Other
Detailed Placers can now also be added to explore other Detailed
Placement algorithms in VTR.