[AP][PartialLegalizer] Added Bi-Partitioning Spreader
Bi-partitioning spreading is a technique where minimum spanning windows
are formed around overfilled regions of the device such that the
capacity of the bins within the window is larger than the utilization of
the bins within the window. These windows are then successively
partitioned, moving the blocks to the different partitions, until the
window is small enough. Due to its divide and conquer nature, this
algorithm tends to be very fast.

Created a basic version of this algorithm which does not take into
account the different types of blocks being spread. It worked
suprisingly well, giving better QoR and runtime than the original
spreader that I decided to merge this in immediately. In the future,
will add these features.

Kept the old spreader in as a separate Global Placer as it may be
interesting to compare since the old multi-commodity flow-based approach
should, in theory, do a better job spreading (at the expense of
runtime). This old spreader may also be interesting to add into other
stages of the flow. Added tests to keep this spreader from regressing.