[VTR][Util] Added Prefix Sum Class
A prefix sum is a classic data structure for quickly finding the sum of
any sub-region of fixed N-Dimensional array of values. It pre-computes
the partial sums of regions within an array and uses that information
later to save time.

This is currently used in the net_cost_handler to pre-compute number of
inter-die connections.

This class can be useful in other algorithms. For example, in a Partial
Legalizer in an AP flow, it can allow for quick lookups into the
capacity and utilization of regions on the FPGA.

Moved the Prefix Sum class from the net_cost_handler into the VTR Util.