[vtr_flow] Changed Triggers for Second Run
Currently, when running the VPR stage of the vtr_flow script, if you are
routing at a fixed channel width and you have "--route" or "--analysis"
in the command, it will run the VPR stage twice with the exact same
arguments. This is unnecessary and can make some of the runs 2x longer
than they need to be.

The purpose of running the VPR flow twice is that the second time would
load in the temporary files produced by the first run to ensure they are
the same values. It appears as though route and analysis used to be part
of this check; however it is no longer being used. I checked the second
run code and having the --route or --analysis flag set has no affect on
the second run's arguments; so the exact same run is dispatched.

I looked through all the regression tests and only found one testcase
that used either of these options; which may be the reason this was
never caught.

This is useful for the AP flow since, to use the AP flow, we need to
specify the --analytical_place flow which inherently does not perfrom
routing. So we have to add --route to our arguments. I noticed that
these runs took 2x longer than they should and found this issue.