[AP] Fixed Small Bug in Solver and Placer
The solver was reserving too much space for a vector, which was causing
the AP flow to crash on very large circuits. Reduced the size of the
reserve to something more manageable.

The Initial Placer portion of APPack had a small bug where it was
ignoring the flat placement. Rewrote the section slightly to fix this
issue and try to place clusters in better places.