[AP][Legalizer] Added Ability to Generate a Mass Report
While working on the mass abstraction in the partial legalizer of the
globlal placer, found that I needed a lot more information on the device
to be able to debug the mass calculations.

Added a command-line option which will generate a mass report if
requested. This mass report contains useful information on the device,
the netlist, and the mass / capacities computed in the mass calculator.