[Faster Reconstruction]
The legalizations strategy of cluster_legalizer is set to
SKIP_LB_ROUTING for the first passs. After first pass, each
checked and illegal ones destroyed and their molecules are
passed to second pass.

The legalization strategy is set to FULL after first pass
in that commit.

Be careful with the place where you compress the legalizer
and extract the atom lookup for placement.