[RouterLookahead] Router Lookahead Report
Added an option to generate a report on the router lookahead. This
reports profiles the router lookahead by performing a set of trial
routes from source RR nodes to all target RR nodes and comparing the
estimated cost from the router lookahead to the actual cost of the path.

Notably, this report tries to show the mean squared error of the router
lookahead and the maximum overestimation of the router lookahead.

The mean squared error is a measure of how accurate the router lookahead
is. The more accurate the router lookahead is, the faster the path
search will be (with less loss in quality). The max overestimation is a
measure of the admissibility of the router lookahead heuristic. The more
the router lookahead overestimates, the worse the quality of the path
found (theoretically).