[Tatum][Parse] Fixed Extraneous Warning With get_clocks
The get_clocks command is used in an SDC file to reference a set of
clocks by name using a regex string. The code to do this tries to
produce a warning if get_clocks is used on a regex string and no clocks
could be found. The issue is that the code to do this was mistakenly
producing this warning for each clock in the circuit. For example, if we
had {clk1, clk2, clk3} and we wanted to do "get_clocks {clk3}", we will
get two warnings since clk1 and clk2 did not match.

Fixed this by moving the warning out of one loop nest.