[Reducing memory usage]
Changed the way of processing mainly in the first pass. In this
version, they are first sorted by external pin number and then
grouped into tiles. Then processed similar to naive by tiles but
this one still checks the compatibility and capacity. Cleaning
after each tile clusters created in that pass. Then, checking
each cluster legality after they are created in the first
neighbour pass. Second is already done with FULL strategy. The
memory usage is decreased by nearly by %35. We lost runtime but it
seems near 1-2%.

TODO for that one: I added an if in neighbour pass that checks if
the molecule is already clustered before checking its compatibility.
This works now but ideally my algorithm should not try to add an
already clustered molecule to another one. When removed, it failed
on vtr_chain_largest for only LU32PEEng.v.