[AP] Illegal Cluster Handling and Partial Placement Verification
This is the version that results are presented in 2 May
Friday meeting. The memory improved version is presented with master
merged. We have things to try on top of that version.

Handling the illegal clusters right away. Also added a 0.5f buffer
into fixed blocks in partial placement verification to pass the
assert. This should be handled more explicitely later.