[AP] Reconstruction Stat Reporting Fix
The displacement for each atom was being calculated as distance
to the head of the tile. Now it considers the offset of the tile
as well. That affects the percent of atoms displaced, max atom
displacemetn and average atom displacement.

Did no touched the cluster error for now. This can also be
rewritten. Instead of being calculated solely on cluster location
and centroid, we can try to use cluster information.