[Infra] Updated Install Packages Script For Backwards Compatibility
The install_apt_packages.sh script is no longer backward compatible with
older versions of Ubuntu due to the dependency on clang-format-18.

Added an if statement to check if the distribution can support
clang-format-18 and only installing it if it can.

Added this script to the CI build process so it can always be tested
within the CI to prevent future regression.