[FASM] Fixed Bug With Wire Creation
Found a bug within FASM's wire generation where it uses the index of the
output pin to create the wire instead of the index of the input pin.

This stemmed from some confusing code which both verified that the wire
was being used and returning the first valid pin. It just so happens
that it checked the outputs first and returned the output pin instead.

Cleaned up the code and added more error checking to prevent issues like
this in the future.