[CI] Added Serial Execution Engine Test
Since the CI always installs oneTBB and the execution engine is set to
auto, I found that the CI always tested with the tbb execution engine.
Some users may not have oneTBB installed for one reason or another and
we need to ensure that VTR always builds.

Added a CI test which sets the parallel execution engine to serial for
Tatum and VPR.