[CI] Removed Redundant Warning Test
Prior to the updates to the CI to make all regression tests warning
clean, there was another warning test which was not as comprehensive as
the tests we have now.

Since this test was superceded, removed it from the CI. The CMAKE param
that enabled it was also used and replaced with a more comprehensive
CMAKE option.