[AP][Timing] Used Flat Placement Info to Compute Setup Criticalities
When timing analysis was turned on for AP, we originally only used the
pre-cluster timing analyzer which was very high-level and innacurate. It
practically just counted the number of hops between launch and capture
registers to approximate criticality.

Improved this by using flat placement information provided by AP.

During global placement, the criticality of all edges are recomputed
using the upper bound solution from the prior iteration of GP. The place
delay model from the placement flow was used to get an mostly-accurate
delay estimation for distances between tiles. The slacks computed each
GP iteration are used to update the net weights between iterations to
better optimize CPD and sTNS.

This improved estimation of setup slacks is then passed into the full
legalizer, which it is then used by the packer to better pack critical
atoms together.

This change required some changes to the APNetlist. Notably, we need all
atom nets to be located somewhere in the AP netlist such that their
delays can be calculated properly. Instead of removing nets we do not
care about for AP, marked them as ignored.