[AP][Solver] Per-Connection Timing Optimization
The B2B solver originally used a simple net weighting model to
incorperate timing into the objective. This works well, but a more
direct approach to timing should provide a better solution.

Added extra edges to the B2B formulation which encode per-connection
delays into the objective.

Simply, these edges incentivice the solver to shrink the distance
between driver and sinks where shrinking that distance would improve the
delay the most (focusing more on currently critical edges).

To encode this objective, I compute the instantaneuous derivative of the
delay of an connection with respect to the distance (normalized such
that the units are not in seconds) and use that as a weight for a pin to
pin connection between each driver and its sinks.