[AP] Optimized Primitive Vector Class
The primitive vector class was assumed to be quite sparse, and as such
used an unordered map as its internal data structure.

Found through experimentation that most of the time in the partial
legalizer was being spent in the operations of the primitive vector
class. Also, while improving the mass abstraction, I found a need to
separate logical models from the dimensions of the primitive vector to
allow multiple models to point to the same dimension in the primitive
vector.

This PR kills two birds with one stone by turning the unordered map into
a VTR vector map and creating a new PrimitiveVectorDim which can allow
the models to be separate from the dimensions.

Future PRs will make use of this feature to improve the mass
abstraction.