[CI] Added Quick Titanium S10 Tests
The titanium benchmarks were not being tested by the CI. Added the
Titanium benchmarks which could be run in under around 2 hours to
NightlyTest7.

5 circuits in this benchmark set currently fail through VTR. The
failures are mainly in the initial placer, which is struggling to create
an initial placement when logical blocks can be placed into different
physical block types which are constrained resources.