[Build] Resolved Most IPO Warnings
VTR is built with IPO by default. IPO tends to reveal false warnings in
the compiler. This was causing many warnings when VTR was built by
default.

Most of these warnings were coming from libarchfpga due to how it
allocates arrays. Suppressed these warnings when IPO is turned on.

There are still link-time warnings coming from CapNProto; however, since
this library is external, it is a harder to supress these warnings. We
could suppress the warnings by globally setting the link option to
ignore the warning, but that may be overkill.