[AP][GP] Added Post-Global Placement Wirelength Estimation
After a recent change where we are ignoring high-fanout nets in the
global placer, this caused the high-fanout nets to also be ignored
during wirlength estimation in the global placement.

This wirelength is used in the global placer to decide when to stop and
to print logs to the user. I would like to continue to ignore these nets
when computing the objective of the global placer since it is the
wirelength of the nets which are not ignored; however, for the user, I
would like them to see a more accurate estimation of the wirelength.

Created a function which tries to estimate the wirelength of the global
placement solution assuming all blocks can be placed exactly where there
flat placement wants them to be placed.